# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Comptes Bancaires",
        "description": "API REST pour la gestion des comptes bancaires, utilisateurs et transactions",
        "contact": {
            "email": "contact@banque.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/annaSock/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes actifs et valides",
                "operationId": "5ddb28854cdd71d2897a05d8cfcc1769",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes avec pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Rôle non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rôle non autorisé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte bancaire",
                "description": "Permet de créer un compte bancaire pour un nouveau client ou un client existant avec validation complète des données.",
                "operationId": "e2c5a6265f53c4ad00eb34f1dc46536a",
                "requestBody": {
                    "description": "Données pour créer un compte. Deux cas possibles : avec un client existant (id) ou en créant un nouveau client.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "soldeInitial",
                                    "devise",
                                    "client"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "cheque",
                                            "epargne"
                                        ],
                                        "example": "cheque"
                                    },
                                    "soldeInitial": {
                                        "type": "number",
                                        "minimum": 10000,
                                        "example": 500000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "enum": [
                                            "XOF",
                                            "FCFA"
                                        ],
                                        "example": "XOF"
                                    },
                                    "client": {
                                        "description": "Informations du client. Utilisez 'id' pour un client existant ou les autres champs pour créer un nouveau client.",
                                        "type": "object",
                                        "oneOf": [
                                            {
                                                "title": "Nouveau client",
                                                "required": [
                                                    "titulaire",
                                                    "email",
                                                    "telephone",
                                                    "adresse"
                                                ],
                                                "properties": {
                                                    "titulaire": {
                                                        "description": "Nom du titulaire",
                                                        "type": "string",
                                                        "example": "Cheikh Sy"
                                                    },
                                                    "nci": {
                                                        "description": "Numéro de carte d'identité nationale",
                                                        "type": "string",
                                                        "example": "1234567890123A"
                                                    },
                                                    "email": {
                                                        "description": "Adresse email",
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "cheikh.sy@example.com"
                                                    },
                                                    "telephone": {
                                                        "description": "Numéro de téléphone",
                                                        "type": "string",
                                                        "example": "+221771234567"
                                                    },
                                                    "adresse": {
                                                        "description": "Adresse",
                                                        "type": "string",
                                                        "example": "Dakar, Sénégal"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            {
                                                "title": "Client existant",
                                                "required": [
                                                    "id"
                                                ],
                                                "properties": {
                                                    "id": {
                                                        "description": "ID du client existant",
                                                        "type": "integer",
                                                        "example": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "titulaire": "Le nom du titulaire est requis",
                                                "soldeInitial": "Le solde initial doit être supérieur à 0"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation métier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/annaSock/v1/comptes/{compte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer un compte spécifique",
                "description": "Permet de récupérer un compte spécifique (chèque ou épargne) quel que soit son statut (actif, bloqué, fermé). Utilise le Route Model Binding et la validation.",
                "operationId": "9a0d50158cd30796c041a1e6235f2a03",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "UUID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte récupéré avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comptes/{compte}": {
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte",
                "operationId": "497a1b9d261322d3077bf6d5581302e9",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte",
                "operationId": "a7891fca561d7991f2804d600f869cc6",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Compte supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://example.com/api/comptes?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://example.com/api/comptes?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://example.com/api/comptes?page=1",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://example.com/api/comptes?page=3",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token Bearer pour l'authentification",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Gestion des comptes bancaires"
        }
    ]
}